+++ src/main/java/roomescape/repository/RoomescapeLockRepository.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
 import org.springframework.stereotype.Repository;
+import roomescape.external.lock.RoomescapeLock;
 
 @Repository
 public interface RoomescapeLockRepository extends JpaRepository<RoomescapeLock, String> {
